requirements.txt
boto3>=1.26.0
botocore>=1.29.0
Pillow>=9.0.0
Flask>=2.0.0